GT4Py programs as DaCe SDFGConvertibles
